+++ modules/__init__.py
@@ -10,3 +10,4 @@ from .load import (
 )
 from .plotting import DataPlotter
 from .features import FeatureFactory as FeatureFactory
+from .augment import DataAugmentation as Augmenter